% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featrank_ranger.R
\name{featrank_ranger}
\alias{featrank_ranger}
\title{Screen features via a fast implementation of Random Forest}
\usage{
featrank_ranger(
  Y,
  X,
  family,
  obsWeights = NULL,
  ties_method = "last",
  nTree = 1000,
  mTry = ifelse(family$family == "gaussian", floor(sqrt(ncol(X))),
    max(floor(ncol(X)/3), 1)),
  nodeSize = ifelse(family$family == "gaussian", 5, 1),
  importanceType = c("permutation", "impurity"),
  scalePermutationImportance = TRUE,
  probabilityTrees = FALSE,
  numThreads = 1,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{importanceType}{Importance type. \code{"permutation"} (default) indicates
mean decrease in accuracy (for \code{binomial()} family) or percent increase
in mean squared error (for \code{gaussian()} family) when comparing
predictions using the original variable versus a permuted version of the
variable (column of \code{X}). \code{"impurity"} indicates increase in
node purity achieved by splitting on that column of \code{X} (for
\code{binomial()} family, measured by Gini index; for \code{gaussian()},
measured by variance of the responses). See
\code{\link[ranger]{ranger}} for more details.}

\item{scalePermutationImportance}{Scale permutation importance by standard
error. Ignored if \code{importanceType = "impurity"}. See
\code{\link[ranger]{ranger}} for more details.}

\item{probabilityTrees}{Logical. If family is \code{binomial()} and
\code{probabilityTrees} is FALSE (the default), classification trees are
grown. If family is \code{binomial()} and
\code{probabilityTrees} is TRUE (the default), probability trees are
grown (Malley et al., 2012). Ignored if family is \code{gaussian()}, for
which regression trees are always grown. See \code{\link[ranger]{ranger}}
for more details.}

\item{numThreads}{Number of threads. Default: 1.}
}
\description{
Speed up \code{\link[SuperLearner]{screen.randomForest}} or
\code{\link{screen.randomForest.imp}}. Uses the
\code{\link[FSelector]{cutoff}} selectors.
}
\references{
\url{http://dx.doi.org/10.18637/jss.v077.i01}
\url{http://dx.doi.org/10.1023/A:1010933404324}
\url{http://dx.doi.org/10.3414/ME00-01-0052}
}
